+++ src/Error/Error.php
@@ -30,11 +30,8 @@ use function iterator_to_array;
  * Class extends standard PHP `\Exception`, so all standard methods of base `\Exception` class
  * are available in addition to those listed below.
  */
-class Error extends Exception implements JsonSerializable, ClientAware
+class Error extends Exception implements JsonSerializable, ClientAware, ProvidesExtensions
 {
-    const CATEGORY_GRAPHQL  = 'graphql';
-    const CATEGORY_INTERNAL = 'internal';
-
     /**
      * Lazily initialized.
      *
@@ -70,8 +67,6 @@ class Error extends Exception implements JsonSerializable, ClientAware
 
     private bool $isClientSafe;
 
-    protected string $category;
-
     /** @var array<string, mixed>|null */
     protected ?array $extensions;
 
@@ -94,34 +89,31 @@ class Error extends Exception implements JsonSerializable, ClientAware
 
         // Compute list of blame nodes.
         if ($nodes instanceof Traversable) {
-            $nodes = iterator_to_array($nodes);
-        } elseif ($nodes !== null && ! is_array($nodes)) {
-            $nodes = [$nodes];
+            $this->nodes = iterator_to_array($nodes);
+        } elseif (is_array($nodes)) {
+            $this->nodes = $nodes;
+        } elseif ($nodes !== null) {
+            $this->nodes = [$nodes];
         }
 
-        $this->nodes      = $nodes;
-        $this->source     = $source;
-        $this->positions  = $positions;
-        $this->path       = $path;
-        $this->extensions = is_array($extensions) && count($extensions) > 0
-            ? $extensions
-            : ($previous instanceof self
-                ? $previous->extensions
-                : []
-            );
+        $this->source    = $source;
+        $this->positions = $positions;
+        $this->path      = $path;
+
+        if (count($extensions) > 0) {
+            $this->extensions = $extensions;
+        } elseif ($previous instanceof ProvidesExtensions) {
+            $this->extensions = $previous->getExtensions();
+        } else {
+            $this->extensions = [];
+        }
 
         if ($previous instanceof ClientAware) {
             $this->isClientSafe = $previous->isClientSafe();
-            $cat                = $previous->getCategory();
-            $this->category     = $cat === '' || $cat === null
-                ? self::CATEGORY_INTERNAL
-                : $cat;
         } elseif ($previous !== null) {
             $this->isClientSafe = false;
-            $this->category     = self::CATEGORY_INTERNAL;
         } else {
             $this->isClientSafe = true;
-            $this->category     = self::CATEGORY_GRAPHQL;
         }
     }
 
@@ -201,11 +193,6 @@ class Error extends Exception implements JsonSerializable, ClientAware
     public function isClientSafe(): bool
     {
         return $this->isClientSafe;
-    }
-
-    public function getCategory(): string
-    {
-        return $this->category;
     }
 
     public function getSource(): ?Source