+++ src/Error/FormattedError.php
@@ -37,6 +37,13 @@ use function strlen;
  * This class is used for [default error formatting](error-handling.md).
  * It converts PHP exceptions to [spec-compliant errors](https://facebook.github.io/graphql/#sec-Errors)
  * and provides tools for error debugging.
+ *
+ * @phpstan-type FormattedErrorArray array{
+ *  message: string,
+ *  locations?: array<int, array{line: int, column: int}>,
+ *  path?: array<int, int|string>,
+ *  extensions?: array<string, mixed>,
+ * }
  */
 class FormattedError
 {
@@ -161,7 +168,7 @@ class FormattedError
      *
      * For a list of available debug flags @see \GraphQL\Error\DebugFlag constants.
      *
-     * @return array<string, mixed>
+     * @return FormattedErrorArray
      *
      * @api
      */
@@ -169,21 +176,11 @@ class FormattedError
     {
         $internalErrorMessage ??= self::$internalErrorMessage;
 
-        if ($exception instanceof ClientAware) {
-            $formattedError = [
-                'message'  => $exception->isClientSafe() ? $exception->getMessage() : $internalErrorMessage,
-                'extensions' => [
-                    'category' => $exception->getCategory(),
-                ],
-            ];
-        } else {
-            $formattedError = [
-                'message'  => $internalErrorMessage,
-                'extensions' => [
-                    'category' => Error::CATEGORY_INTERNAL,
-                ],
-            ];
-        }
+        $message = $exception instanceof ClientAware && $exception->isClientSafe()
+            ? $exception->getMessage()
+            : $internalErrorMessage;
+
+        $formattedError = ['message' => $message];
 
         if ($exception instanceof Error) {
             $locations = [];
@@ -198,10 +195,12 @@ class FormattedError
             if ($exception->path !== null && count($exception->path) > 0) {
                 $formattedError['path'] = $exception->path;
             }
+        }
 
+        if ($exception instanceof ProvidesExtensions) {
             $extensions = $exception->getExtensions();
-            if (is_array($extensions)) {
-                $formattedError['extensions'] = $extensions + $formattedError['extensions'];
+            if (count($extensions) > 0) {
+                $formattedError['extensions'] = $extensions;
             }
         }
 
@@ -213,14 +212,15 @@ class FormattedError
     }
 
     /**
-     * Decorates spec-compliant $formattedError with debug entries according to $debug flags
-     * (@see \GraphQL\Error\DebugFlag for available flags)
+     * Decorates spec-compliant $formattedError with debug entries according to $debug flags.
      *
-     * @param mixed[] $formattedError
+     * phpcs:disable SlevomatCodingStandard.Commenting.DocCommentSpacing.IncorrectAnnotationsGroup
+     * @param int                 $debugFlag      For available flags @see \GraphQL\Error\DebugFlag
      *
-     * @return mixed[]
+     * phpcs:disable SlevomatCodingStandard.TypeHints.ParameterTypeHint.MissingTraversableTypeHintSpecification
+     * @param FormattedErrorArray $formattedError
      *
-     * @throws Throwable
+     * @return FormattedErrorArray
      */
     public static function addDebugEntries(array $formattedError, Throwable $e, int $debugFlag): array
     {