+++ tests/Executor/AbstractTest.php
@@ -426,14 +426,13 @@ class AbstractTest extends TestCase
             'errors' => [
                 [
                     'message'      => 'Internal server error',
+                    'locations'    => [['line' => 1, 'column' => 3]],
+                    'path'         => ['foo'],
                     'debugMessage' =>
                         'Abstract type FooInterface must resolve to an Object type at ' .
                         'runtime for field Query.foo with value "dummy", received "[]". ' .
                         'Either the FooInterface type should provide a "resolveType" ' .
                         'function or each possible type should provide an "isTypeOf" function.',
-                    'locations'    => [['line' => 1, 'column' => 3]],
-                    'path'         => ['foo'],
-                    'extensions'   => ['category' => 'internal'],
                 ],
             ],
         ];